requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9